# Remove sync.Pool from internal/repository