… internal/repository/pool.go